♻️ Chartコンポーネントのリファクタリング
