+++ src/app/(pages)/(contents)/dashboard/components/Chart/index.ts
@@ -1,1 +1,1 @@
-export * from "./MyChart";
+export * from "./Chart";
